# src/Agentd.Web/ClientApps/shared/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Agentd.Host | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/setup/database": {
      "get": {
        "tags": [
          "Agentd.Host"
        ],
        "operationId": "GetSetupDatabase",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseStepVm"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Agentd.Host"
        ],
        "operationId": "SaveSetupDatabase",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/DatabaseRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveResultVm"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/setup/database/test": {
      "post": {
        "tags": [
          "Agentd.Host"
        ],
        "operationId": "TestSetupDatabase",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/DatabaseRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StepCheckVm"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/setup/azure-devops": {
      "get": {
        "tags": [
          "Agentd.Host"
        ],
        "operationId": "GetSetupAzureDevOps",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AzureDevOpsStepVm"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Agentd.Host"
        ],
        "operationId": "SaveSetupAzureDevOps",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/AzureDevOpsRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveResultVm"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/setup/azure-devops/test": {
      "post": {
        "tags": [
          "Agentd.Host"
        ],
        "operationId": "TestSetupAzureDevOps",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/AzureDevOpsRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StepCheckVm"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AzureDevOpsRequest": {
        "required": [
          "organization",
          "project",
          "auth",
          "pat"
        ],
        "type": "object",
        "properties": {
          "organization": {
            "type": [
              "null",
              "string"
            ]
          },
          "project": {
            "type": [
              "null",
              "string"
            ]
          },
          "auth": {
            "type": [
              "null",
              "string"
            ]
          },
          "pat": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "\u0060Auth\u0060 is \u0060Pat\u0060 or \u0060AzCli\u0060. An empty \u0060Pat\u0060 keeps the saved one."
      },
      "AzureDevOpsStepVm": {
        "required": [
          "organization",
          "project",
          "auth",
          "pat"
        ],
        "type": "object",
        "properties": {
          "organization": {
            "type": [
              "null",
              "string"
            ]
          },
          "project": {
            "type": [
              "null",
              "string"
            ]
          },
          "auth": {
            "type": "string"
          },
          "pat": {
            "$ref": "#/components/schemas/SecretStatusVm"
          }
        },
        "description": "\u0060Auth\u0060 is \u0060Pat\u0060 or \u0060AzCli\u0060."
      },
...
      "DatabaseRequest": {
        "required": [
          "connectionString"
        ],
        "type": "object",
        "properties": {
          "connectionString": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "The PostgreSQL connection string (a secret). On Test, empty means \u0022the saved one\u0022."
      },
      "DatabaseStepVm": {
        "required": [
          "connectionString"
        ],
        "type": "object",
        "properties": {
          "connectionString": {
            "$ref": "#/components/schemas/SecretStatusVm"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SaveResultVm": {
        "required": [
          "restartRequired"
        ],
        "type": "object",
        "properties": {
          "restartRequired": {
            "type": "boolean"
          }
        },
        "description": "Saved. The daemon reads these settings at start, so they apply after a restart."
      },
      "SecretStatusVm": {
        "required": [
          "set",
          "updatedAt",
          "updatedBy"
        ],
        "type": "object",
        "properties": {
          "set": {
            "type": "boolean"
          },
          "updatedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "updatedBy": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "A secret\u0027s status. Secrets are write-only: the value never leaves the server."
      },
...
      "StepCheckVm": {
        "required": [
          "ok",
          "message",
          "fix"
        ],
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "fix": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "A step\u0027s \u0022Test\u0022 result; string? StepCheckVm.Fix says what to do when it failed."
      },
...
    }
  },
...
}